@@ -24,7 +24,7 @@
         },
         "sort-packages": true,
         "platform": {
-            "php": "8.0.99"
+            "php": "8.1.99"
         }
     },
     "extra": {
@@ -34,7 +34,7 @@
         }
     },
     "require": {
-        "php": "~8.0.0 || ~8.1.0 || ~8.2.0 || ~8.3.0",
+        "php": "~8.1.0 || ~8.2.0 || ~8.3.0",
         "psr/http-factory": "^1.0.2",
         "psr/http-message": "^1.1 || ^2.0"
     },
@@ -48,7 +48,7 @@
         "php-http/psr7-integration-tests": "^1.3",
         "phpunit/phpunit": "^9.5.28",
         "psalm/plugin-phpunit": "^0.18.4",
-        "vimeo/psalm": "^5.9"
+        "vimeo/psalm": "^5.15.0"
     },
     "provide": {
         "psr/http-factory-implementation": "^1.1 || ^2.0",
