@@ -34,7 +34,9 @@ final class ApiAndInternalAnnotationCheck implements Rule
             return [];
         }
 
-        if (str_starts_with($reflection->getName(), 'CuyZ\Valinor\Tests')) {
+        if (str_starts_with($reflection->getName(), 'CuyZ\Valinor\Tests')
+            || str_starts_with($reflection->getName(), 'SimpleNamespace')
+        ) {
             return [];
         }
 
