@@ -357,7 +357,7 @@ class PriorityQueue implements Countable, IteratorAggregate, Serializable
             if (! $this->queue instanceof \SplPriorityQueue) {
                 throw new Exception\DomainException(sprintf(
                     'PriorityQueue expects an internal queue of type SplPriorityQueue; received "%s"',
-                    $this->queue::class
+                    $queue::class
                 ));
             }
         }
