@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Laminas\Stdlib;
 
+use ReturnTypeWillChange;
 use Serializable;
 use UnexpectedValueException;
 
@@ -37,17 +38,18 @@ class SplPriorityQueue extends \SplPriorityQueue implements Serializable
      * Utilizes {@var $serial} to ensure that values of equal priority are
      * emitted in the same order in which they are inserted.
      *
-     * @param  TValue    $datum
+     * @param  TValue    $value
      * @param  TPriority $priority
      * @return void
      */
-    public function insert($datum, $priority)
+    #[ReturnTypeWillChange] // Inherited return type should be bool
+    public function insert($value, $priority)
     {
         if (! is_array($priority)) {
             $priority = [$priority, $this->serial--];
         }
 
-        parent::insert($datum, $priority);
+        parent::insert($value, $priority);
     }
 
     /**
