@@ -9,9 +9,23 @@
  */
 namespace SebastianBergmann\Diff;
 
-final class Diff
+use ArrayIterator;
+use IteratorAggregate;
+use Traversable;
+
+/**
+ * @template-implements IteratorAggregate<int, Chunk>
+ */
+final class Diff implements IteratorAggregate
 {
+    /**
+     * @psalm-var non-empty-string
+     */
     private string $from;
+
+    /**
+     * @psalm-var non-empty-string
+     */
     private string $to;
 
     /**
@@ -20,6 +34,8 @@ final class Diff
     private array $chunks;
 
     /**
+     * @psalm-param non-empty-string $from
+     * @psalm-param non-empty-string $to
      * @psalm-param list<Chunk> $chunks
      */
     public function __construct(string $from, string $to, array $chunks = [])
@@ -29,12 +45,18 @@ final class Diff
         $this->chunks = $chunks;
     }
 
-    public function getFrom(): string
+    /**
+     * @psalm-return non-empty-string
+     */
+    public function from(): string
     {
         return $this->from;
     }
 
-    public function getTo(): string
+    /**
+     * @psalm-return non-empty-string
+     */
+    public function to(): string
     {
         return $this->to;
     }
@@ -42,7 +64,7 @@ final class Diff
     /**
      * @psalm-return list<Chunk>
      */
-    public function getChunks(): array
+    public function chunks(): array
     {
         return $this->chunks;
     }
@@ -53,5 +75,40 @@ final class Diff
     public function setChunks(array $chunks): void
     {
         $this->chunks = $chunks;
+    }
+
+    /**
+     * @psalm-return non-empty-string
+     *
+     * @deprecated
+     */
+    public function getFrom(): string
+    {
+        return $this->from;
+    }
+
+    /**
+     * @psalm-return non-empty-string
+     *
+     * @deprecated
+     */
+    public function getTo(): string
+    {
+        return $this->to;
+    }
+
+    /**
+     * @psalm-return list<Chunk>
+     *
+     * @deprecated
+     */
+    public function getChunks(): array
+    {
+        return $this->chunks;
+    }
+
+    public function getIterator(): Traversable
+    {
+        return new ArrayIterator($this->chunks);
     }
 }
